Fix MD5 portability on arm64 and add Phase 1 EIBStdLib unit tests

+++ EIBStdLib/include/MD5.h
@@ -2,6 +2,7 @@
 #define __MD5_HEADER__
 
 #include "CString.h"
+#include <cstdint>
 
 typedef unsigned char *POINTER;
 
@@ -10,8 +11,8 @@ typedef unsigned char *POINTER;
  */
 typedef struct
 {
-	unsigned long int state[4];   	      /* state (ABCD) */
-	unsigned long int count[2]; 	      /* number of bits, modulo 2^64 (lsb first) */
+	uint32_t state[4];   	      /* state (ABCD), MD5 uses 32-bit words */
+	uint32_t count[2]; 	      /* number of bits, modulo 2^64 (lsb first) */
 	unsigned char buffer[64];	          /* input buffer */
 } MD5_CTX;
 
@@ -23,9 +24,9 @@ class EIB_STD_EXPORT MD5
 {
 private:
 
-	void MD5Transform (unsigned long int state[4], unsigned char block[64]);
-	void Encode (unsigned char*, unsigned long int*, unsigned int);
-	void Decode (unsigned long int*, unsigned char*, unsigned int);
+	void MD5Transform (uint32_t state[4], unsigned char block[64]);
+	void Encode (unsigned char*, uint32_t*, unsigned int);
+	void Decode (uint32_t*, unsigned char*, unsigned int);
 	void MD5_memcpy (POINTER, POINTER, unsigned int);
 	void MD5_memset (POINTER, int, unsigned int);
