+++ EIBStdLib/src/MD5.cpp
@@ -62,23 +62,23 @@ FF, GG, HH, and II transformations for rounds 1, 2, 3, and 4.
 Rotation is separate from addition to prevent recomputation.
 */
 #define FF(a, b, c, d, x, s, ac) { \
- (a) += F ((b), (c), (d)) + (x) + (unsigned long int)(ac); \
+ (a) += F ((b), (c), (d)) + (x) + static_cast<uint32_t>(ac); \
  (a) = ROTATE_LEFT ((a), (s)); \
  (a) += (b); \
   }
 
 #define GG(a, b, c, d, x, s, ac) { \
- (a) += G ((b), (c), (d)) + (x) + (unsigned long int)(ac); \
+ (a) += G ((b), (c), (d)) + (x) + static_cast<uint32_t>(ac); \
  (a) = ROTATE_LEFT ((a), (s)); \
  (a) += (b); \
   }
 #define HH(a, b, c, d, x, s, ac) { \
- (a) += H ((b), (c), (d)) + (x) + (unsigned long int)(ac); \
+ (a) += H ((b), (c), (d)) + (x) + static_cast<uint32_t>(ac); \
  (a) = ROTATE_LEFT ((a), (s)); \
  (a) += (b); \
   }
 #define II(a, b, c, d, x, s, ac) { \
- (a) += I ((b), (c), (d)) + (x) + (unsigned long int)(ac); \
+ (a) += I ((b), (c), (d)) + (x) + static_cast<uint32_t>(ac); \
  (a) = ROTATE_LEFT ((a), (s)); \
  (a) += (b); \
   }
@@ -106,11 +106,11 @@ void MD5::MD5Update (MD5_CTX *context, unsigned char *input, unsigned int inputL
 	  index = (unsigned int)((context->count[0] >> 3) & 0x3F);
 
 	  /* Update number of bits */
-	  if ( (context->count[0] += ((unsigned long int)inputLen << 3))
-	       < ((unsigned long int)inputLen << 3))
-		context->count[1]++;
+  if ( (context->count[0] += (static_cast<uint32_t>(inputLen) << 3))
+       < (static_cast<uint32_t>(inputLen) << 3))
+	context->count[1]++;
 
-	  context->count[1] += ((unsigned long int)inputLen >> 29);
+  context->count[1] += (static_cast<uint32_t>(inputLen) >> 29);
 	  partLen = 64 - index;
 
 	  /*
@@ -169,9 +169,9 @@ void MD5::MD5Final (unsigned char digest[16], MD5_CTX *context)
 /*
  * MD5 basic transformation. Transforms state based on block.
  */
-void MD5::MD5Transform (unsigned long int state[4], unsigned char block[64])
+void MD5::MD5Transform (uint32_t state[4], unsigned char block[64])
 {
-	unsigned long int a = state[0], b = state[1], c = state[2], d = state[3], x[16];
+	uint32_t a = state[0], b = state[1], c = state[2], d = state[3], x[16];
 
 	Decode (x, block, 64);
 
@@ -263,7 +263,7 @@ void MD5::MD5Transform (unsigned long int state[4], unsigned char block[64])
  * Encodes input (unsigned long int) into output (unsigned char). Assumes len is
  * a multiple of 4.
  */
-void MD5::Encode (unsigned char *output, unsigned long int *input, unsigned int len)
+void MD5::Encode (unsigned char *output, uint32_t *input, unsigned int len)
 {
 	unsigned int i, j;
 
@@ -279,15 +279,15 @@ void MD5::Encode (unsigned char *output, unsigned long int *input, unsigned int 
  * Decodes input (unsigned char) into output (unsigned long int). Assumes len is
  * a multiple of 4.
  */
-void MD5::Decode (unsigned long int *output, unsigned char *input, unsigned int len)
+void MD5::Decode (uint32_t *output, unsigned char *input, unsigned int len)
 {
 	  unsigned int i, j;
 
 	  for (i = 0, j = 0; j < len; i++, j += 4)
-		 output[i] = ((unsigned long int)input[j]) |
-			     (((unsigned long int)input[j+1]) << 8) |
-			     (((unsigned long int)input[j+2]) << 16) |
-			     (((unsigned long int)input[j+3]) << 24);
+	 output[i] = (static_cast<uint32_t>(input[j])) |
+		     (static_cast<uint32_t>(input[j+1]) << 8) |
+		     (static_cast<uint32_t>(input[j+2]) << 16) |
+		     (static_cast<uint32_t>(input[j+3]) << 24);
 }
 
 /*
